Develop

+++ routes/web.php
@@ -24,30 +24,44 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\EventPositionPresetController;
 use App\Http\Controllers\EventFieldController;
 use App\Http\Controllers\EventController;
+use App\Http\Controllers\ManageEventController;
+use App\Http\Controllers\VisitFacilityController;
 use App\Mail\TrainingAssignmentCreated;
-use App\EventPositionAssignmentController;
 use App\Http\Controllers\EventPositionController;
 use App\Livewire\EventRegistration;
 
+# Homepage
 Route::get('/', [HomeController::class, 'index'])->name('home');
+
+# Roster
+Route::get('/roster', [RosterController::class, 'index'])->name('roster.index');
+
+# Visit
+Route::get('/visit', [VisitFacilityController::class, 'index'])->name('visit.index');
+Route::get('/visit/create', [VisitFacilityController::class, 'create'])->middleware('auth')->name('visit.create');
+Route::post('/visit', [VisitFacilityController::class, 'store'])->middleware('auth')->name('visit.store');
 
 # Oauth
 Route::get('/auth/redirect', [VatsimOauthController::class, 'redirect'])->name('auth.redirect');
+Route::get('/login', function() {
+    return redirect()->route('auth.redirect', 301);
+})->name('login');
 Route::get('/auth/callback', [VatsimOauthController::class, 'callback'])->name('auth.callback');
 Route::get('/auth/logout', [VatsimOauthController::class, 'logout'])->name('auth.logout');
 
-Route::resource('users', UserController::class);
-
-Route::resource('users', UserController::class, ['only' => ['edit', 'update']]);
-Route::prefix('users/{user}')->group(function () {
+# Users
+Route::resource('users', UserController::class, ['only' => ['show', 'edit', 'update']]);
+Route::prefix('users/{user}')->group(function() {
     Route::get('/', [UserController::class, 'show'])->name('users.show');
     Route::get('training-tickets', [UserController::class, 'trainingTickets'])->name('users.show.training-tickets');
     Route::get('training-assignments', [UserController::class, 'trainingAssignments'])->name('users.show.training-assignments');
     Route::get('solo-certs', [UserController::class, 'soloCerts'])->name('users.show.solo-certs');
 });
 
+# Staff Directory
 Route::get('/staff', [StaffController::class, 'index'])->name('staff.index');
 
+# Training Assignment Creation; TODO: make store
 Route::post('training-assignment/create', [TrainingAssignmentController::class, 'create'])->middleware('auth')->name('training-assignment.create');
 Route::prefix('events')->name('events.')->group(function () {
     Route::get('/', [EventController::class, 'index'])->name('index');
@@ -55,19 +69,33 @@ Route::prefix('events')->name('events.')->group(function () {
 });
 Route::post('/events/{event}/request-position', [EventRegistration::class, 'store'])->middleware('auth')->name('events.request-position.store');
 
-Route::prefix('admin')->middleware('permission:view dashboard')->group(function () {
+# Admin Routes
+Route::prefix('admin')->middleware('permission:view dashboard')->group(function() {
+    # Dashboard
     Route::get('/', [DashboardController::class, 'index'])->name('admin.index');
-    Route::get('users', [UserManagementController::class, 'index'])->name('manage-users.index');
 
-    Route::middleware('permission:manage statistics prefixes')->group(function () {
+    # User Management
+    Route::get('users', [UserManagementController::class, 'index'])->name('manage-users.index');
+    Route::middleware('permission:manage visiting controllers')->group(function() {
+        Route::get('visit-requests/{visitRequest}', [VisitFacilityController::class, 'show'])->name('visit.show');
+        Route::get('visit-requests', [VisitFacilityController::class, 'manage'])->name('visit.manage');
+        Route::put('visit-requests/{visitRequest}', [VisitFacilityController::class, 'update'])->name('visit.update');
+        Route::put('visit-requests/{visitRequest}/approve', [VisitFacilityController::class, 'approve'])->name('visit.approve');
+        Route::put('visit-requests/{visitRequest}/deny', [VisitFacilityController::class, 'deny'])->name('visit.deny');
+    });
+
+    # Facilities Dept.
+    Route::middleware('permission:manage statistics prefixes')->group(function() {
         Route::resource('statistics-prefixes', StatisticsPrefixesController::class);
     });
 
-    Route::middleware('permission:view audit logs')->group(function () {
+    # Logs
+    Route::middleware('permission:view audit logs')->group(function() {
         Route::get('logs', [AuditLogController::class, 'index'])->name('logs.index');
     });
 
-    Route::prefix('/training')->middleware('role:training')->group(function () {
+    # Training Dept.
+    Route::prefix('/training')->middleware('role:training')->group(function() {
         Route::resource('tickets', TrainingTicketController::class)->names('training-tickets');
         Route::resource('assignments', TrainingAssignmentController::class, ['only' => ['update', 'edit', 'index']])->names('training-assignments');
         Route::resource('solo-certs', SoloCertController::class, ['only' => ['index', 'create', 'update', 'destroy', 'store']])->names('solo-certs');
@@ -76,20 +104,20 @@ Route::prefix('admin')->middleware('permission:view dashboard')->group(function 
         Route::delete('assignments', [TrainingAssignmentController::class, 'destroy'])->name('training-assignments.destroy'); //id sent in payload
     });
 
-    Route::prefix('events')->middleware('permission:manage events')->name('admin.events.')->group(function () {
-        Route::resource('event-fields', EventFieldController::class)->names('event-fields');
-        Route::resource('position-presets', EventPositionPresetController::class)->names('position-presets');
-        Route::get('/', [EventController::class, 'manage'])->name('index');
-        Route::get('create', [EventController::class, 'create'])->name('create');
-        Route::post('/', [EventController::class, 'store'])->name('store');
-        Route::get('{event}/edit', [EventController::class, 'edit'])->name('edit');
-        Route::put('{event}', [EventController::class, 'update'])->name('update');
-        Route::delete('{event}', [EventController::class, 'destroy'])->name('destroy');
+    # Events Dept.
+    Route::prefix('/events')->middleware('permission:manage events')->group(function () {
+        Route::resource('event-fields', EventFieldController::class)->names('admin.events.event-fields');
+        Route::resource('position-presets', EventPositionPresetController::class)->names('admin.events.position-presets');
+        Route::get('/', [EventController::class, 'manage'])->name('admin.events.index');
+        Route::get('create', [EventController::class, 'create'])->name('admin.events.create');
+        Route::post('/', [EventController::class, 'store'])->name('admin.events.store');
+        Route::get('{event}/edit', [EventController::class, 'edit'])->name('admin.events.edit');
+        Route::put('{event}', [EventController::class, 'update'])->name('admin.events.update');
+        Route::delete('{event}', [EventController::class, 'destroy'])->name('admin.events.destroy');
     });
 });
 
-Route::get('/roster', [RosterController::class, 'index'])->name('roster.index');
-
+# Dev Only Routes
 if (App::environment('development', 'local')) {
     Route::get('/sync', function () {
         SyncRoster::dispatch();
